requirements.txt
streamlit==1.12.0
pandas==1.4.3
numpy==1.22.4
regex==2022.8.17
contractions==0.1.72
scikit-learn==1.1.2
Pillow==8.3.1
mbti_data
mbti_ava